출석시 Session 모델에 담긴 Code로 출석하도록 수정
- RemoteConfig에 존재하는 공통의 출석코드가 아닌 각 세션에 존재하는 Code를 Input과 비교하여 출석하도록 수정

+++ domain/src/main/java/com/yapp/domain/repository/SessionRepository.kt
@@ -4,6 +4,6 @@ import com.yapp.domain.model.Session
 
 interface SessionRepository {
     suspend fun setSession(session: Session): Result<Unit>
-    suspend fun getSession(id: Long): Result<Session?>
+    suspend fun getSession(id: Int): Result<Session?>
     suspend fun getAllSession(): Result<List<Session>>
 }
